+++ eslint.config.js
@@ -4,10 +4,10 @@ export default [
   ...componentHookPlugin.configs.basic,
   ...componentHookPlugin.configs.vue,
   ...componentHookPlugin.configs.react,
-  ...componentHookPlugin.configs.markdown,
   componentHookPlugin.configs.prettier,
   ...componentHookPlugin.configs.sonarjs,
   componentHookPlugin.configs.security,
+  ...componentHookPlugin.configs.markdown,
   {
     files: ['**/__tests__/unit/**/*.test.[jt]s?(x)'],
     ...componentHookPlugin.configs['testing-library/vue'],